+++ tests/integration/pure_cairo_vm/cheatcodes/prank/prank_cairo_cheatcode_test.py
@@ -1,0 +1,18 @@
+from pathlib import Path
+
+from tests.integration.conftest import assert_cairo_test_cases
+from tests.integration.pure_cairo_vm.conftest import RunCairoTestRunnerFixture
+
+
+async def test_prank_cheatcode(run_cairo_test_runner: RunCairoTestRunnerFixture):
+    testing_summary = await run_cairo_test_runner(
+        Path(__file__).parent / "prank_test.cairo",
+        test_cases=["test_not_fails_when_pranked"],
+    )
+
+    assert_cairo_test_cases(
+        testing_summary,
+        expected_passed_test_cases_names=[
+            "test_not_fails_when_pranked",
+        ],
+    )
